feat: Create new `Notes` table
A polymorphic Notes table was needed to store all unstructured notes.

This commit creates the necessary polymorphic Notes table. Its purpose is to store all unstructured notes, categorized by a note_type. It should be used when a record might need more than one note,
when the notes need to be categorized, or when you need the ability to
search across all notes in the system. This is different from a dedicated
notes field on a specific table, which should be used to store a single,
intrinsic attribute of the record itself.