refactor: update lexicon category names and terms associated with the StatusHistory table.
`well_status` is defined as a lexicon category, but it should be a lexicon value associated with the `status_type` category.

 The `status_type` category does not exist, but it should.

The terms assigned to the `well_status` category should actually be assigned to the `status_value` category.

Renamed `well_status` category to `status_value`.
Added new `status_type` category and related terms.
Updated term definitions.