refactor: Add ThingResponseForContact schema and update ContactResponse
If a full ThingResponse is used in ContactResponse, it can lead to
performance issues due to lazy loading of unneeded fields. To preclude
this from happening, a new ThingResponseForContact schema has been created
since all that is needed in this context are the id and name of the Things
related to a Contact.